+++ src/canread_lpc1768.cpp
@@ -9,9 +9,9 @@ CanMessage receiveCanMessage(CanBus* bus) {
 
     CanMessage result = {message.id, 0};
     result.data = message.dataA[0];
-    result.data |= (message.dataA[1] << 8);
-    result.data |= (message.dataA[2] << 16);
-    result.data |= (message.dataA[3] << 24);
+    result.data |= (((uint64_t)message.dataA[1]) << 8);
+    result.data |= (((uint64_t)message.dataA[2]) << 16);
+    result.data |= (((uint64_t)message.dataA[3]) << 24);
     result.data |= (((uint64_t)message.dataB[0]) << 32);
     result.data |= (((uint64_t)message.dataB[1]) << 40);
     result.data |= (((uint64_t)message.dataB[2]) << 48);
